+++ app/dashboard/templates/onepager/base.html
@@ -89,5 +89,6 @@
       }, 1000);
     </script>
     {% endif %}
+    {% include 'shared/messages.html' %}
   </body>
 </html>
