+++ app/dashboard/templates/onepager/receive.html
@@ -17,17 +17,9 @@
 {% endcomment %}
 {% load i18n static %}
 {% block 'scripts' %}
-<script src="{% static "onepager/js/quarantine/ethereumjs-accounts.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/web3.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/hooked-web3-provider.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/tx.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/index.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/tokens.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/pre.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/receive.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/confetti.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/web3alert.js" %}"></script>
-<script src="{% static "onepager/js/quarantine/lightwallet.min.js" %}"></script>
+<script src="{% static "v2/js/shared.js" %}"></script>
+<script src="{% static "onepager/js/receive.js" %}"></script>
+<script src="{% static "onepager/js/confetti.js" %}"></script>
 
 {% endblock %}
 <!-- Main -->
@@ -43,47 +35,35 @@
       </a>
     </span>
   </header>
-  <div id="loading">
-    <h1>{% trans "Loading" %}</h1>
-    <img src="{% static "onepager/images/loading.gif" %}" style="max-width: 70px; max-height: 70px;">
-    <p id="loading_txt">{% trans "waiting for web3..." %}</p>
+  <div id="receove_eth" {% if tip.receive_txid %} style="display:none;" {%endif%} >
+    <form method="POST">
+      <h1 id="oneh1">{% trans "You've got" %} {{tip.amount}} {{tip.tokenName}} </h1>
+      <div>
+        {% trans "Forward it to a permanent address:" %}
+        <br>
+        <input type="hidden" name="network" id="network" value="">
+        <input type="text" name="forwarding_address" id="forwarding_address" placeholder="0x0" {% if disable_inputs %}disabled=disabled{%endif%}>
+      </div>
+      <div>
+        <br>
+        <input id="receive" type='submit' class="button button--primary" value="{% trans "Claim" %} {{tip.tokenName}} 🙌" {% if disable_inputs %}disabled=disabled{%endif%} >
+        <br>
+      </div>
+    </form>
   </div>
-  <div id="step_zero" style="display:none;">
-    <h1 id="zeroh1">{% trans "You've Got ETH!" %}</h1>
-    <p>{% trans "To claim it, please install MetaMask and refresh the page." %}</p>
-    <a href="https://metamask.io/?utm_source=gitcoin.co&utm_medium=referral" id="install" class="button button--primary">{% trans "Install Metamask" %} <span class="emoji">🐬</span></a>
-    <br>
-    <br>
-    <h2>{% trans "What is Metamask?" %}</h2>
-    <iframe width="100%" src="https://www.youtube.com/embed/6Gf_kRE4MJU" frameborder="0" allowfullscreen></iframe>
-  </div>
-  <div id="send_eth" style="display:none;">
-    <h1 id="oneh1">{% trans "Receive Ethereum Now." %}</h1>
-    <div>
-      <input type="hidden" placeholder="idx_address" id="idx_address" value="">
-    </div>
-    <div>
-      <input type="hidden" placeholder="private_key" id="private_key" value="">
-    </div>
-    <div>
-      {% trans "Forward it to a permanent address:" %}
-      <br>
-      <input type="text" placeholder="forwarding_address" id="forwarding_address" value="0x0">
-    </div>
-    <div>
-      <br>
-      <a href="#" id="receive" class="button button--primary">{% trans "Claim" %} <span id="tokenName">Eth</span> {% trans "Tip" %} <span class="emoji">🙌</span></a>
-      <br>
-    </div>
-  </div>
-  <div id="send_eth_done" style="display:none;">
+  <div id="receove_eth_done" {% if not tip.receive_txid %} style="display:none;" {%endif%}>
     <h1>{% trans "Congrats!" %}</h1>
-    <p>{% blocktrans %}You've received Ethereum via transaction <span id="trans_id"></span>.{% endblocktrans %}</p>
+    <p>You've received {{tip.amount}} {{tip.tokenName}} via transaction <span id="trans_id">{{tip.receive_txid}}</span>.</p>
   </div>
+{% if tip.receive_txid and not disable_inputs %}
+  <script>
+    setTimeout(function(){
+      startConfetti();
+    }, 1000);
+  </script>
+{% endif %}
 </section>
 <script>
-  document.pk = '{{tip.metadata.priv_key}}';
-  document.txid = '{{tip.txid}}';
   document.network = '{{tip.network}}';
 </script>
 {% endblock %}
