+++ app/dashboard/views.py
@@ -39,7 +39,7 @@ from django.views.decorators.http import require_GET, require_POST
 
 from app.utils import ellipses, sync_profile
 from avatar.utils import get_avatar_context
-from dashboard.utils import generate_pub_priv_keypair
+from dashboard.utils import generate_pub_priv_keypair, get_web3
 from gas.utils import conf_time_spread, gas_advisories, recommend_min_gas_price_to_confirm_in_time
 from github.utils import (
     get_auth_url, get_github_emails, get_github_primary_email, get_github_user_data, is_github_token_valid,
@@ -71,17 +71,6 @@ confirm_time_minutes_target = 4
 
 # web3.py instance
 w3 = Web3(HTTPProvider(settings.WEB3_HTTP_PROVIDER))
-
-
-def send_tip(request):
-    """Handle the first stage of sending a tip."""
-    params = {
-        'issueURL': request.GET.get('source'),
-        'title': _('Send Tip'),
-        'class': 'send',
-    }
-
-    return TemplateResponse(request, 'onepager/send1.html', params)
 
 
 def record_user_action(user, event_name, instance):
@@ -466,245 +455,6 @@ def uninterested(request, bounty_id, profile_id):
         'success': True,
         'msg': _("You've stopped working on this, thanks for letting us know."),
     })
-
-
-@csrf_exempt
-@ratelimit(key='ip', rate='2/m', method=ratelimit.UNSAFE, block=True)
-def receive_tip_legacy(request):
-    """Receive a tip."""
-    if request.body:
-        status = 'OK'
-        message = _('Tip has been received')
-        params = json.loads(request.body)
-
-        # db mutations
-        try:
-            tip = Tip.objects.get(txid=params['txid'])
-            tip.receive_address = params['receive_address']
-            tip.receive_txid = params['receive_txid']
-            tip.received_on = timezone.now()
-            tip.save()
-            record_user_action(tip.username, 'receive_tip', tip)
-            record_tip_activity(tip, tip.username, 'receive_tip')
-        except Exception as e:
-            status = 'error'
-            message = str(e)
-
-        # http response
-        response = {
-            'status': status,
-            'message': message,
-        }
-
-        return JsonResponse(response)
-
-    params = {
-        'issueURL': request.GET.get('source'),
-        'class': 'receive',
-        'title': _('Receive Tip'),
-        'gas_price': round(recommend_min_gas_price_to_confirm_in_time(confirm_time_minutes_target), 1),
-    }
-
-    return TemplateResponse(request, 'onepager/receive_legacy.html', params)
-
-
-@csrf_exempt
-@ratelimit(key='ip', rate='2/m', method=ratelimit.UNSAFE, block=True)
-def receive_tip_v2(request, pk, txid, network):
-
-    tip = Tip.objects.get(web3_type='v2', metadata__priv_key=pk, txid=txid, network=network)
-
-    """Receive a tip."""
-    if request.body:
-        status = 'OK'
-        message = _('Tip has been received')
-        params = json.loads(request.body)
-
-        # db mutations
-        try:
-            tip = Tip.objects.get(txid=params['txid'])
-            tip.receive_address = params['receive_address']
-            tip.receive_txid = params['receive_txid']
-            tip.received_on = timezone.now()
-            tip.save()
-            record_user_action(tip.username, 'receive_tip', tip)
-            record_tip_activity(tip, tip.username, 'receive_tip')
-        except Exception as e:
-            status = 'error'
-            message = str(e)
-
-        # http response
-        response = {
-            'status': status,
-            'message': message,
-        }
-
-        return JsonResponse(response)
-
-    params = {
-        'issueURL': request.GET.get('source'),
-        'class': 'receive',
-        'title': _('Receive Tip'),
-        'gas_price': round(recommend_min_gas_price_to_confirm_in_time(confirm_time_minutes_target), 1),
-        'tip': tip,
-    }
-
-    return TemplateResponse(request, 'onepager/receive.html', params)
-
-
-@csrf_exempt
-@ratelimit(key='ip', rate='5/m', method=ratelimit.UNSAFE, block=True)
-def send_tip_4(request):
-    """Handle the fourth stage of sending a tip (the POST)
-
-    Returns:
-        JsonResponse: response with success state.
-
-    """
-    response = {
-        'status': 'OK',
-        'message': _('Tip Sent'),
-    }
-
-    is_user_authenticated = request.user.is_authenticated
-    from_username = request.user.username if is_user_authenticated else ''
-    primary_from_email = request.user.email if is_user_authenticated else ''
-    access_token = request.user.profile.get_access_token() if is_user_authenticated else ''
-    to_emails = []
-
-    params = json.loads(request.body)
-    txid = params['txid']
-    destinationAccount = params['destinationAccount']
-    tip = Tip.objects.get(metadata__address=destinationAccount)
-    is_authenticated_for_this_via_login = (tip.from_username and tip.from_username == from_username)
-    is_authenticated_for_this_via_ip = tip.ip == get_ip(request)
-    is_authed = is_authenticated_for_this_via_ip or is_authenticated_for_this_via_login
-    if not is_authed:
-        response = {
-            'status': 'error',
-            'message': _('Permission Denied'),
-        }
-        return JsonResponse(response)
-
-    # db mutations
-    tip.txid = txid
-    tip.save()
-
-    # notifications
-    maybe_market_tip_to_github(tip)
-    maybe_market_tip_to_slack(tip, 'new_tip')
-    maybe_market_tip_to_email(tip, to_emails)
-    record_user_action(tip.username, 'send_tip', tip)
-    record_tip_activity(tip, tip.from_name, 'new_tip')
-
-    return JsonResponse(response)
-
-
-
-@csrf_exempt
-@ratelimit(key='ip', rate='5/m', method=ratelimit.UNSAFE, block=True)
-def send_tip_3(request):
-    """Handle the third stage of sending a tip (the POST)
-
-    Returns:
-        JsonResponse: response with success state.
-
-    """
-    response = {
-        'status': 'OK',
-        'message': _('Tip Created'),
-    }
-
-    is_user_authenticated = request.user.is_authenticated
-    from_username = request.user.username if is_user_authenticated else ''
-    primary_from_email = request.user.email if is_user_authenticated else ''
-    access_token = request.user.profile.get_access_token() if is_user_authenticated else ''
-    to_emails = []
-
-    params = json.loads(request.body)
-
-    to_username = params['username'].lstrip('@')
-    try:
-        to_profile = Profile.objects.get(handle__iexact=to_username)
-    except Profile.MultipleObjectsReturned:
-        to_profile = Profile.objects.filter(handle__iexact=to_username).first()
-    except Profile.DoesNotExist:
-        to_profile = None
-    if to_profile:
-        if to_profile.email:
-            to_emails.append(to_profile.email)
-        if to_profile.github_access_token:
-            to_emails = get_github_emails(to_profile.github_access_token)
-
-    if params.get('email'):
-        to_emails.append(params['email'])
-
-    # If no primary email in session, try the POST data. If none, fetch from GH.
-    if params.get('fromEmail'):
-        primary_from_email = params['fromEmail']
-    elif access_token and not primary_from_email:
-        primary_from_email = get_github_primary_email(access_token)
-
-    to_emails = list(set(to_emails))
-    expires_date = timezone.now() + timezone.timedelta(seconds=params['expires_date'])
-    priv_key, pub_key, address = generate_pub_priv_keypair()
-
-    # db mutations
-    tip = Tip.objects.create(
-        emails=to_emails,
-        tokenName=params['tokenName'],
-        amount=params['amount'],
-        comments_priv=params['comments_priv'],
-        comments_public=params['comments_public'],
-        ip=get_ip(request),
-        expires_date=expires_date,
-        github_url=params['github_url'],
-        from_name=params['from_name'],
-        from_email=params['from_email'],
-        from_username=from_username,
-        username=params['username'],
-        network=params['network'],
-        tokenAddress=params['tokenAddress'],
-        from_address=params['from_address'],
-        metadata={
-            'priv_key': priv_key,
-            'pub_key': pub_key,
-            'address': address,
-        }
-    )
-    response['payload'] = {
-        'address': address,
-    }
-    return JsonResponse(response)
-
-
-@ratelimit(key='ip', rate='5/m', method=ratelimit.UNSAFE, block=True)
-def send_tip_2(request):
-    """Handle the second stage of sending a tip.
-
-    TODO:
-        * Convert this view-based logic to a django form.
-
-    Returns:
-        JsonResponse: If submitting tip, return response with success state.
-        TemplateResponse: Render the submission form.
-
-    """
-
-    is_user_authenticated = request.user.is_authenticated
-    from_username = request.user.username if is_user_authenticated else ''
-    primary_from_email = request.user.email if is_user_authenticated else ''
-
-    params = {
-        'issueURL': request.GET.get('source'),
-        'class': 'send2',
-        'title': _('Send Tip'),
-        'recommend_gas_price': recommend_min_gas_price_to_confirm_in_time(confirm_time_minutes_target),
-        'from_email': primary_from_email,
-        'from_handle': from_username,
-    }
-
-    return TemplateResponse(request, 'onepager/send2.html', params)
 
 
 def onboard(request, flow):
